Support for mangled names, minor fix for import table
For re targeting mangled names in .def file, target name also has to be
mangled in order for the keep the exported names same and the solution
in the given link applied heere by removing 'extern "C"' for mangled
symbols. That way thee global functions defined in the cpp has mangled
names and it works for what we need.

Link to explanation about mangled names:
https://floodyberry.wordpress.com/2008/09/08/generating-dll-wrappers/

Import table has to have mangled symbol name for GetProcAddress to work
properly.